adicionado cometa

+++ generator/main.cpp
@@ -76,7 +76,7 @@ int main(int argc, char *argv[]) {
         if(argc != 8) {
             cout << "Wrong number of arguments!" << endl;
         } else {
-            Torus tor(atof(argv[2]), atoi(argv[3]), atoi(argv[4]), atoi(argv[5]), atoi(argv[6]));
+            Torus tor(atof(argv[2]), atof(argv[3]), atof(argv[4]), atoi(argv[5]), atoi(argv[6]));
             tor.dump_new(redir(argv[7]));
         }
     } else {
